+++ modules/ai-openai/nodejs/CHANGELOG.md
@@ -1,12 +1,12 @@
 # @telorun/ai-openai
 
-## 1.0.0
+## 0.3.0
 
 ### Patch Changes
 
 - Updated dependencies [ae0bf77]
-  - @telorun/sdk@1.0.0
-  - @telorun/ai@1.0.0
+  - @telorun/sdk@0.13.0
+  - @telorun/ai@0.3.0
 
 ## 0.2.1