+++ modules/console/nodejs/CHANGELOG.md
@@ -1,11 +1,11 @@
 # @telorun/console
 
-## 1.0.0
+## 0.7.0
 
 ### Patch Changes
 
 - Updated dependencies [ae0bf77]
-  - @telorun/sdk@1.0.0
+  - @telorun/sdk@0.13.0
 
 ## 0.6.0