+++ modules/lambda/nodejs/CHANGELOG.md
@@ -1,12 +1,12 @@
 # @telorun/lambda
 
-## 1.0.0
+## 0.4.0
 
 ### Patch Changes
 
 - Updated dependencies [ae0bf77]
-  - @telorun/sdk@1.0.0
-  - @telorun/http-dispatch@1.0.0
+  - @telorun/sdk@0.13.0
+  - @telorun/http-dispatch@0.4.0
 
 ## 0.3.1