+++ modules/mcp-server/nodejs/CHANGELOG.md
@@ -1,11 +1,11 @@
 # @telorun/mcp-server
 
-## 1.0.0
+## 0.6.0
 
 ### Patch Changes
 
 - Updated dependencies [ae0bf77]
-  - @telorun/sdk@1.0.0
+  - @telorun/sdk@0.13.0
 
 ## 0.5.1