+++ modules/test/nodejs/CHANGELOG.md
@@ -1,12 +1,12 @@
 # @telorun/test
 
-## 1.0.0
+## 0.4.0
 
 ### Patch Changes
 
 - Updated dependencies [ae0bf77]
-  - @telorun/sdk@1.0.0
-  - @telorun/kernel@1.0.0
+  - @telorun/sdk@0.13.0
+  - @telorun/kernel@0.15.0
 
 ## 0.1.2