+++ sdk/nodejs/CHANGELOG.md
@@ -1,6 +1,6 @@
 # @telorun/sdk
 
-## 1.0.0
+## 0.13.0
 
 ### Minor Changes